[API] move CampaignToJoin to read models
